Remove "Invalid FAT type" error message in try_fs_geometry()
This function is called in a loop for each of the 3 FatType to find out
which of the type can be used give some parameters, and it is designed
to fail for some of the FatType, so it doesn't make much sense to print
this error message.

Without this patch, formatting a small partition (which will select a
Fat12 type) always print the "Invalid FAT type" error twice, even
though it finishes sucessfully.